restructuring app.jsx into smaller, readable files
