+++ src/hooks/useTheme.js
@@ -1,0 +1,33 @@
+// src/hooks/useTheme.js
+import { useState, useEffect } from "react";
+
+export function useTheme() {
+  const [theme, setTheme] = useState(() => {
+    if (
+      typeof window !== "undefined" &&
+      window.localStorage &&
+      typeof window.localStorage.getItem === "function"
+    ) {
+      return localStorage.getItem("theme") || "dark"; // default: dark
+    }
+    return "dark";
+  });
+
+  useEffect(() => {
+    // Matches original: toggle "light-theme" class on body
+    if (theme === "light") {
+      document.body.classList.add("light-theme");
+    } else {
+      document.body.classList.remove("light-theme");
+    }
+
+    if (typeof localStorage !== "undefined" && localStorage.setItem) {
+      localStorage.setItem("theme", theme);
+    }
+  }, [theme]);
+
+  const toggleTheme = () =>
+    setTheme((prev) => (prev === "dark" ? "light" : "dark"));
+
+  return { theme, toggleTheme };
+}
